chore(code-garden): extract AC parsing module from feature-task main.rs [2026-W32]
Non-functional change from repo audit 2026-W32.
Finding: [Medium] Feature-task workflow remains a 7,120-line Rust entrypoint.

Move AC parsing functions (extract_ac_section, parse_evidence, strip_trailing_evidence,
parse_ac_line, verify_pr_acs_failures, task_description_acs, unchecked_task_ac_labels,
ac_labels_in_pr_body, ac_labels_needing_new_pr, task_ac_vs_open_pr_failures) plus the
Evidence enum and AcEvidence struct into a new ac_parsing module. main.rs drops from
7120 to ~6280 lines and gains a single `mod ac_parsing;` declaration. Evidence/AC
parsing logic now has a focused module boundary with its own unit tests in-place.

Behaviour is unchanged: all 197 tests pass, clippy is clean with -D warnings.
