+++ docs/repo-audits/2026-W32.md
@@ -38,7 +38,7 @@
 - **Why:** A typo or partial deployment can change policy while the service remains healthy; the warning is easy to miss and callers receive a successful response.
 - **Severity:** Medium.
 
-**[Medium] Feature-task workflow remains a 7,120-line Rust entrypoint.**
+**[Medium] Feature-task workflow remains a 7,120-line Rust entrypoint.** ✅ [PR #TBD](https://github.com/Stoffer-Industries/sindustries/pull/TBD) (AC parsing stage extracted — entrypoint ~940 lines smaller; further stages pending)
 
 - **Where:** `agents/workflows/feature-task/src/main.rs:1-80` declares only `mod analytics`; the file is 7,120 lines and contains stage/parsing functions through at least `:3547`.
 - **Why:** Loading, approval-source selection, PR inspection, AC parsing, delivery verification, analytics, and post-merge behavior have a large shared regression surface. Extracting stage modules would improve review and focused testing.
